Added Eigen dependency for use with the physics engine.

+++ simulator/api/phys.cpp
@@ -43,7 +43,9 @@ double Phys::scale = 1;
 void Phys::seedMersenne()
 {
     // rng.seed(std::random_device()());
-    unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
+    unsigned long seed =
+            static_cast<unsigned long>(std::chrono::system_clock::now().time_since_epoch().count());
+
     std::random_device rd;
     //std::seed_seq seed2{seed, rd(), rd()};
     //rng.seed(seed);
